+++ client/src/components/content/clientTable/ClientTable.jsx
@@ -1,5 +1,6 @@
 import React, { useEffect, useState } from 'react'
 import { useDispatch, useSelector } from "react-redux";
+import Swal from 'sweetalert2'
 import { deleteClient, getClients, putClient,} from '../../../redux/action/clients';
 
 const ClientTable = () => {
@@ -10,8 +11,26 @@ const ClientTable = () => {
     }, [])
 
     const deleteC = (id) => {
-        dispatch(deleteClient(id))
-        window.location.reload()
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+        }).then((result) => {
+            if (result.isConfirmed) {
+                dispatch(deleteClient(id))
+                Swal.fire(
+                    'Deleted!',
+                    'Your file has been deleted.',
+                    'success'
+                )
+                window.location.reload()
+            }
+        })
+        
     }
 
     function onChange(e) {
@@ -32,9 +51,28 @@ const ClientTable = () => {
                 },
             };
             dispatch(putClient(client.id, clientSend))
-            window.location.reload()
+            Swal.fire({
+                title: 'Promoted!!!',
+                icon: 'success',
+                confirmButtonColor: '#3085d6',
+                confirmButtonText: 'Ok'
+              }).then((result) => {
+                if (result.isConfirmed) {
+                    window.location.reload()
+                }
+              })
         }
-        else alert("this account is suspended, please renovate it and try again")
+        else {
+            Swal.fire({
+                title: 'this account is suspended, please renovate it and try again',
+                showClass: {
+                  popup: 'animate__animated animate__fadeInDown'
+                },
+                hideClass: {
+                  popup: 'animate__animated animate__fadeOutUp'
+                }
+              })
+            }
     }
 
     const changeStatus = (client) => {
@@ -45,7 +83,16 @@ const ClientTable = () => {
                 },
             };
             dispatch(putClient(client.id, clientSend))
-            window.location.reload()
+            Swal.fire({
+                title: 'Banned!!!',
+                icon: 'success',
+                confirmButtonColor: '#3085d6',
+                confirmButtonText: 'Ok'
+              }).then((result) => {
+                if (result.isConfirmed) {
+                    window.location.reload()
+                }
+              })
         } else {
             const clientSend = {
                 clientModified: {
@@ -53,7 +100,16 @@ const ClientTable = () => {
                 },
             };
             dispatch(putClient(client.id, clientSend))
-            window.location.reload()
+            Swal.fire({
+                title: 'Renovated!!!',
+                icon: 'success',
+                confirmButtonColor: '#3085d6',
+                confirmButtonText: 'Ok'
+              }).then((result) => {
+                if (result.isConfirmed) {
+                    window.location.reload()
+                }
+              })
         }
     }
